requirements.txt
#absl-py==1.0.0
#certifi==2022.9.24
datasets
debugpy
evaluate
flatbuffers
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
#h5py
huggingface-hub==0.29
#imageio==2.19.0
ipywidgets
#Markdown==3.3.6
matplotlib==3.9
#numexpr==2.8.1
numpy==2.2
#oauthlib==3.2.0
opencv-python==4.11.0.86
#opt-einsum==3.3.0
#os-sys==2.1
packaging==21.3
pandas
Pillow==9.0.1
#pluggy==1.0.0
protobuf
#py==1.11.0
pyparsing==3.0.4
pytest==7.1.2
#python-dateutil==2.8.2
requests==2.32
#requests-oauthlib==1.3.1
rsa==4.8
scikit-image
scikit-learn==1.6
scipy
tqdm==4.67
torch==2.6
transformers==4.49
#typeguard==2.13.3
#typing_extensions==4.3.0
#urllib3==1.26.9
#virtualenv==20.4.2
 